+++ rekorder/lib/track/manager.py
@@ -30,6 +30,7 @@ class TrackManager:
     else:
       raise Exception("Unsupported mode [{}]".format(self.mode))
 
+    self.current_track = None
     self.reset()
 
   @staticmethod
@@ -41,6 +42,7 @@ class TrackManager:
     return self
 
   def __next__(self):
+    self._check_state()
     was = self.current_track.title if self.current_track else ''
     if self._index < self._max:
       self.current_track = self._tracks[self._index]
@@ -50,8 +52,19 @@ class TrackManager:
       raise StopIteration
 
   def set_track(self, new_track):
+    self._check_state()
     while self.current_track and self.current_track.title != new_track:
       next(self)
+
+  def sub_track_begin(self, new_sub_track):
+    sub_track = Track(index=self.current_track.index * 10, title=new_sub_track, parent=self.current_track)
+    # self.current_track.sub_track = sub_track
+    self.current_track = sub_track
+    return self.current_track
+
+  def sub_track_conclude(self, new_sub_track):
+    self.current_track = self.current_track.parent
+    return self.current_track
 
   @property
   def tracks(self):
@@ -62,7 +75,14 @@ class TrackManager:
     return self.__next__()
 
   def reset(self):
+    self._check_state()
     self.current_track = None
     self._index = 0
     self._max = len(self._tracks)
     return self.next
+
+  def _check_state(self):
+    if not self.current_track:
+      return
+    if self.current_track.parent:
+      raise Exception("Illegal state.")